New style to document on FreightDetails and option to delete document
