+++ mobile-app/components/pages/manager/fleets/[fleet_id]/trucks/[truck_id]/freights/[freight_id]/details/FreightDetailsScreen.tsx
@@ -1,17 +1,14 @@
-import { Button, makeStyles } from "@rneui/themed";
-import * as FileSystem from "expo-file-system";
-import * as MediaLibrary from "expo-media-library";
+import { makeStyles } from "@rneui/themed";
+
 import { t } from "i18next";
 
 import { useAsyncData } from "@/hooks/use-async-data";
 import { FreightDetailsWithExpensesSchema } from "@/schemas";
-import { api, Log } from "@/services";
-import { uploadMultipartImage } from "@/services/api/uploaders/imageUploaderMultipart";
-import PhotoService from "@/services/photo-service";
-import { colors, iconSize } from "@/services/theme/constants";
+import { api } from "@/services";
 
 import { Header, ScreenWrapper } from "@/components/ui";
 
+import { DocumentDetails } from "./_DocumentDetails";
 import { ExpensesHisotry } from "./_expenses-history/_ExpensesHistory";
 import { FreightDetails } from "./_FreightDetails";
 
@@ -40,62 +37,6 @@ export function FreightDetailsScreen({
     };
   });
 
-  const handleUploadDocument = async () => {
-    try {
-      const image = await PhotoService.pick();
-
-      if (image) {
-        const resizedImage = await PhotoService.resize(image.uri, 256, 256);
-
-        const uploadSuccess = await uploadMultipartImage({
-          url:
-            api().defaults.baseURL +
-            `/fleets/${fleetId}/trucks/${truckId}/freights/${freightId}/document`,
-          fieldName: "document",
-          imageUri: resizedImage.uri,
-        });
-
-        if (uploadSuccess) {
-          await refresh();
-        }
-      }
-    } catch (e: any) {
-      Log.e(e);
-    }
-  };
-
-  const handleDownloadDocument = async () => {
-    try {
-      const url =
-        api().defaults.baseURL +
-        `/fleets/${fleetId}/trucks/${truckId}/freights/${freightId}/download`;
-
-      const extension = freight.document_path?.split(".").pop();
-
-      const tempFileUri =
-        FileSystem.cacheDirectory +
-        `Frete_${freight.contractor_name + freight.id}.${extension}`;
-
-      // Baixar o arquivo para cache local
-      const { uri } = await FileSystem.downloadAsync(url, tempFileUri);
-
-      // Solicitar permissão, apenas para imagens (photo)
-      const permissionResponse = await MediaLibrary.requestPermissionsAsync(
-        false,
-        ["photo"]
-      );
-
-      if (permissionResponse.status !== MediaLibrary.PermissionStatus.GRANTED) {
-        return;
-      }
-
-      // Salvar no álbum — método simplificado: saveToLibraryAsync
-      await MediaLibrary.saveToLibraryAsync(uri);
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
   if (loading) return loading;
 
   return (
@@ -104,35 +45,7 @@ export function FreightDetailsScreen({
 
       <FreightDetails freight={freight} />
 
-      {freight.document_path ? (
-        <Button
-          type="outline"
-          title={t("buttons.download-image")}
-          size="sm"
-          iconRight
-          icon={{
-            name: "download",
-            type: "matterial-community",
-            color: colors.primary,
-            size: iconSize.sm,
-          }}
-          onPress={handleDownloadDocument}
-        />
-      ) : (
-        <Button
-          type="outline"
-          title={t("buttons.upload-image")}
-          size="sm"
-          iconRight
-          icon={{
-            name: "upload",
-            type: "matterial-community",
-            color: colors.primary,
-            size: iconSize.sm,
-          }}
-          onPress={handleUploadDocument}
-        />
-      )}
+      <DocumentDetails freight={freight} refresh={refresh} />
 
       <ExpensesHisotry
         expenses={freight.expenses}
